version 1.0.0.7 of ecclem
README.md:
- add links to ICY plugins website
EC-CLEM:
- importRoiPointsFromFile : Add an option to have x,y in physical units,
with an offset of 1 (slice 1 is 0 in Icy) + modify the type of ROI to
ROI 3D point  + add the name to easily see the center of gravity of the
3D point.
- ec-clem : fixed a bug on "show Roi on source image" by adding the
correct pixel size in 3D. The bug was only hapening when the pixel size
were different in source and target image.
- minor ecclem:  change the type conversion after 2D transform
(deprecated method in icy)
-minor ec-clem : remove an unused variable+ update sequence
concatenation according to icy api last updates.

EC-CLEM AUTOFINDER:
- fixed a bug on checkbox "export to ec-clem" for preparing data for
further registration or analysis in EC-CLEM
OrthoViewerRoi:
- no change